api/requirements.txt
fastapi[standard]
motor[srv]
python-decouple
PyJWT
bcrypt
langchain>=0.1.0
langgraph>=0.0.19
langchain_core>=0.1.0
langchain_chroma
langchain_community
langchain_ollama
langchain_huggingface
langchain_qdrant
pypdf
qdrant_client
pydantic>=2.0.0
typing-extensions>=4.5.0
